fix(fitting): drop explicit float dtypes — JAX defaults respect jax_enable_x64
User caught: the previous 4-cell §7.1 matrix showed ~zero throughput
difference between float32 and float64 on RTX Pro 6000 (where f64 should
be ~8-32× slower). Root cause: explicit dtype=jnp.float32 in:

  src/prolix/fitting/init.py   (load_params_init_json — TRAINABLE params)
  src/prolix/fitting/batched.py (stack_molecules — padded arrays)

locked all the bonded params to float32 regardless of the --float64 flag
and the jax.config.update('jax_enable_x64', True) toggle. Compute was
silently f32 throughout in the "f64" cells.

Empirically verified (2026-05-21):

  jnp.zeros((3,))     → float32 when x64 off,  float64 when x64 on
  jnp.array([1.0])    → float32 when x64 off,  float64 when x64 on
  jnp.asarray(np_f64) → float32 when x64 off,  float64 when x64 on

So the cleanest pattern is just: don't specify dtype for float arrays.
JAX's defaults already do the right thing per the global x64 config.
No helper module needed; no jnp.float_ alias (deprecated in NumPy 2.0
anyway); no canonicalize_dtype boilerplate.

Refactored both files to drop ALL `dtype=jnp.float32` / `dtype=jnp.float_`
on jnp.zeros/array calls. Int (int32) and bool dtypes preserved
unchanged (they don't have x64 ambiguity).

Verified post-fix:
  - 17/17 fitting tests still pass
  - load_params_init_json + stack_molecules return float32 with x64 off
    and float64 with x64 on (correct propagation)

Running cluster array (job 14191171) is using the OLD code with the
f32 lock — the 12 f64 cells in that array are invalid. The 12 f32
cells are valid (behavior identical with or without this fix when
x64 is off). Will resubmit just the 12 f64 cells after the running
array drains.

+++ src/prolix/fitting/batched.py
@@ -87,11 +87,11 @@ def stack_molecules(
     n_torsion_terms = params_list[0].n_torsion_terms if params_list[0].n_torsions > 0 else 1
 
     # Initialize padded arrays
-    k_bond_stacked = jnp.zeros((B, max_n_bonds), dtype=jnp.float32)
-    r0_stacked = jnp.zeros((B, max_n_bonds), dtype=jnp.float32)
-    k_theta_stacked = jnp.zeros((B, max_n_angles), dtype=jnp.float32)
-    theta0_rad_stacked = jnp.zeros((B, max_n_angles), dtype=jnp.float32)
-    k_phi_stacked = jnp.zeros((B, max_n_torsions, n_torsion_terms), dtype=jnp.float32)
+    k_bond_stacked = jnp.zeros((B, max_n_bonds))
+    r0_stacked = jnp.zeros((B, max_n_bonds))
+    k_theta_stacked = jnp.zeros((B, max_n_angles))
+    theta0_rad_stacked = jnp.zeros((B, max_n_angles))
+    k_phi_stacked = jnp.zeros((B, max_n_torsions, n_torsion_terms))
 
     # Padded indices: use DISTINCT atom indices so r_ij != 0 for padded bonds.
     # If padded bonds use (0,0), the gradient ∂‖r‖/∂r = r/‖r‖ → 0/0 NaN even
@@ -102,7 +102,7 @@ def stack_molecules(
     angle_idx_stacked = jnp.tile(jnp.array([0, 1, 2], dtype=jnp.int32), (B, max_n_angles, 1))
     torsion_idx_stacked = jnp.tile(jnp.array([0, 1, 2, 3], dtype=jnp.int32), (B, max_n_torsions, 1))
     torsion_periodicity_stacked = jnp.zeros((B, max_n_torsions, n_torsion_terms), dtype=jnp.int32)
-    torsion_phase_rad_stacked = jnp.zeros((B, max_n_torsions, n_torsion_terms), dtype=jnp.float32)
+    torsion_phase_rad_stacked = jnp.zeros((B, max_n_torsions, n_torsion_terms))
 
     bond_mask_stacked = jnp.zeros((B, max_n_bonds), dtype=jnp.bool_)
     angle_mask_stacked = jnp.zeros((B, max_n_angles), dtype=jnp.bool_)
